Finish Trio support, use ThreadWorker base class
All the differences between asyncio/trio are now abstracted into
the ThreadWorker class.
