+++ sqlalchemy_aio/__init__.py
@@ -8,4 +8,4 @@ __all__ = [
 __author__ = 'Frazer McLean <frazer@frazermclean.co.uk>'
 __version__ = '0.12.0'
 __license__ = 'MIT'
-__description__ = 'Asyncio strategy for SQLAlchemy.'
+__description__ = 'Async support for SQLAlchemy.'
