+++ sqlalchemy_aio/trio.py
@@ -1,24 +1,58 @@
-from ._base import AsyncEngine
-from ._worker import ThreadWorker
+import trio
+import threading
+from functools import partial
+
+from trio import Cancelled, RunFinishedError
+from trio.hazmat import Result
+
+from .base import AlreadyQuit, AsyncEngine, ThreadWorker
+
+_STOP = object()
+
+
+class TrioThreadWorker(ThreadWorker):
+    def __init__(self):
+        self._portal = trio.BlockingTrioPortal()
+        self._request_queue = trio.Queue(1)
+        self._response_queue = trio.Queue(1)
+        thread = threading.Thread(target=self.thread_fn, daemon=True)
+        thread.start()
+        self._has_quit = False
+
+    def thread_fn(self):
+        while True:
+            try:
+                request = self._portal.run(self._request_queue.get)
+            except Cancelled:
+                continue
+            except RunFinishedError:
+                break
+
+            if request is not _STOP:
+                response = Result.capture(request)
+                self._portal.run(self._response_queue.put, response)
+            else:
+                self._portal.run(self._response_queue.put, None)
+                break
+
+    async def run(_self, _func, *args, **kwargs):
+        if _self._has_quit:
+            raise AlreadyQuit
+
+        if args or kwargs:
+            _func = partial(_func, *args, **kwargs)
+        await _self._request_queue.put(_func)
+        resp = await _self._response_queue.get()
+        return resp.unwrap()
+
+    async def quit(self):
+        self._has_quit = True
+        await self._request_queue.put(_STOP)
+        await self._response_queue.get()
 
 
 class TrioEngine(AsyncEngine):
     """Mostly like :class:`sqlalchemy.engine.Engine` except some of the methods
     are coroutines."""
-    def __init__(self, pool, dialect, url, logging_name=None, echo=None,
-                 execution_options=None, **kwargs):
-
-        super().__init__(
-            pool, dialect, url, logging_name, echo, execution_options, **kwargs)
-
-        self._engine_worker = None
-
-    async def _run_in_thread(_self, _func, *args, **kwargs):
-        if _self._engine_worker is None:
-            _self._engine_worker = ThreadWorker()
-
-        return await _self._engine_worker.run(_func, *args, **kwargs)
-
-    def _make_connection_thread_fn(self):
-        worker = ThreadWorker()
-        return worker.run, worker.quit
+    def _make_worker(self):
+        return TrioThreadWorker()
