+++ tests/asyncio/test_result_proxy.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import func, select
 from sqlalchemy.schema import CreateTable
 
-from sqlalchemy_aio._base import AsyncResultProxy
+from sqlalchemy_aio.base import AsyncResultProxy
 
 
 @pytest.mark.asyncio
